+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.200.1
 
+- Fix HTML parser to disable media syntax in `sup`, `sub`, and `small` HTML tags.
 - Fix HTML parser to allow `bdo` and `bdi` HTML tags to be nested.
 
 ## 0.200.0